# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Delivery route optimization API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DeliverySessionStatus = typeof DeliverySessionStatus[keyof typeof DeliverySessionStatus];


export const DeliverySessionStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface DeliverySession {
  id: number;
  name: string;
  date: string;
  status: DeliverySessionStatus;
  /** @nullable */
  startAddress?: string | null;
  totalStops?: number;
  completedStops?: number;
  /** @nullable */
  estimatedDurationMinutes?: number | null;
  createdAt: string;
}

export interface DeliverySessionInput {
  name: string;
  date: string;
  startAddress?: string;
}

export type DeliverySessionDetailStatus = typeof DeliverySessionDetailStatus[keyof typeof DeliverySessionDetailStatus];


export const DeliverySessionDetailStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export type StopStatus = typeof StopStatus[keyof typeof StopStatus];


export const StopStatus = {
  pending: 'pending',
  delivered: 'delivered',
  failed: 'failed',
  skipped: 'skipped',
} as const;

export interface Stop {
  id: number;
  sessionId: number;
  address: string;
  /** @nullable */
  recipientName?: string | null;
  /** @nullable */
  items?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  orderIndex: number;
  status: StopStatus;
  /** @nullable */
  estimatedArrival?: string | null;
  createdAt: string;
}

export interface DeliverySessionDetail {
  id: number;
  name: string;
  date: string;
  status: DeliverySessionDetailStatus;
  /** @nullable */
  startAddress?: string | null;
  stops: Stop[];
  /** @nullable */
  estimatedDurationMinutes?: number | null;
  createdAt: string;
}

export interface StopInput {
  address: string;
  recipientName?: string;
  items?: string;
  notes?: string;
  lat?: number;
  lng?: number;
}

export type StopUpdateStatus = typeof StopUpdateStatus[keyof typeof StopUpdateStatus];


export const StopUpdateStatus = {
  pending: 'pending',
  delivered: 'delivered',
  failed: 'failed',
  skipped: 'skipped',
} as const;

export interface StopUpdate {
  address?: string;
  recipientName?: string;
  items?: string;
  notes?: string;
  status?: StopUpdateStatus;
  orderIndex?: number;
}

export interface StopBatch {
  stops: StopInput[];
}

export interface OptimizeRouteInput {
  startLat?: number;
  startLng?: number;
}

export interface Waypoint {
  lat: number;
  lng: number;
}

export interface OptimizedRoute {
  orderedStops: Stop[];
  totalDistanceKm: number;
  estimatedDurationMinutes: number;
  waypoints?: Waypoint[];
}

export interface ExtractAddressesInput {
  imageBase64: string;
  mimeType: string;
}

export interface ExtractedStop {
  address: string;
  /** @nullable */
  recipientName?: string | null;
  /** @nullable */
  items?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface ExtractAddressesResult {
  stops: ExtractedStop[];
  /** @nullable */
  rawText?: string | null;
}

export interface GeocodeResult {
  address: string;
  lat: number;
  lng: number;
  /** @nullable */
  displayName?: string | null;
}

export type GeocodeAddressParams = {
address: string;
};