# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Delivery route optimization API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stops/{stopId}:
    patch:
      operationId: updateStop
      tags: [stops]
      summary: Update a stop (status, items, notes)
      parameters:
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stop"
    delete:
      operationId: deleteStop
      tags: [stops]
      summary: Delete a stop
      parameters:
        - name: stopId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Stop:
      type: object
      required: [id, sessionId, address, orderIndex, status, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        address:
          type: string
        recipientName:
          type: ["string", "null"]
        items:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        orderIndex:
          type: integer
        status:
          type: string
          enum: [pending, delivered, failed, skipped]
        estimatedArrival:
          type: ["string", "null"]
        createdAt:
          type: string
...
    StopUpdate:
      type: object
      properties:
        address:
          type: string
        recipientName:
          type: string
        items:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [pending, delivered, failed, skipped]
        orderIndex:
          type: integer
...