# lib/api-zod/src/generated/types/stop.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Delivery route optimization API
 * OpenAPI spec version: 0.1.0
 */
import type { StopStatus } from './stopStatus';

export interface Stop {
  id: number;
  sessionId: number;
  address: string;
  /** @nullable */
  recipientName?: string | null;
  /** @nullable */
  items?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  orderIndex: number;
  status: StopStatus;
  /** @nullable */
  estimatedArrival?: string | null;
  createdAt: string;
}