# lib/api-zod/src/generated/types/stopUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Delivery route optimization API
 * OpenAPI spec version: 0.1.0
 */
import type { StopUpdateStatus } from './stopUpdateStatus';

export interface StopUpdate {
  address?: string;
  recipientName?: string;
  items?: string;
  notes?: string;
  status?: StopUpdateStatus;
  orderIndex?: number;
}